Creating new helper just filters of rooms.
